SLING-9915 Refactored ServiceGetter.java and made public such that it can be used by AnnotationsProcessor.java
