+++ src/main/java/org/apache/sling/junit/ServiceGetter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.sling.junit.rules;
+package org.apache.sling.junit;
 
 import java.io.Closeable;
 
@@ -27,7 +27,7 @@ import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.util.tracker.ServiceTracker;
 
 /** Implements the logic used to get a service */
-class ServiceGetter<T> implements Closeable {
+public class ServiceGetter<T> implements Closeable {
 
     private final ServiceTracker tracker;
     private final BundleContext bundleContext;
